cli: cfg
This cleans up `main.rs`.
